rename and add docstring
